move upload image modal to another file
