+++ src/styles/post/MakePost.scss
@@ -108,20 +108,3 @@ $icon-color: #787878;
     }
   }
 }
-
-.upload-image-modal {
-  .upload-button-wrapper {
-    display: flex;
-    width: 100%;
-    justify-content: center;
-    
-    .upload-button {
-      width: 100%;
-      background-color: $accent-color;
-      color: white;
-      outline: none;
-      border: 0;
-      cursor: pointer;
-    }
-  }
-}
